requirement.txt
streamlit==1.28.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.0
google-generativeai>=0.3.0
streamlit-extras>=1.0.5